+++ sandbox/factors.py
@@ -45,7 +45,7 @@ class factors:
         if self.svd is not None:
             l,s,r=self.svd # left , singularvals, right
             nf=self.nfactors
-            I=np.identity(nf,np.float)
+            I=np.identity(nf,float)
             suse=I*s[:nf]
             ls = np.dot(l[:,:nf], suse)
             self.gendata=np.dot(ls,r[:nf,:])
@@ -74,10 +74,10 @@ class factors:
                                   s.shape[0],
                                   r.shape[0],
                                   r.shape[1]))
-            print(len(l.astype(np.float).tostring()))
-            out.write(l.astype(np.float).tostring())
-            out.write(s.astype(np.float).tostring())
-            out.write(r.astype(np.float).tostring())
+            print(len(l.astype(float).tostring()))
+            out.write(l.astype(float).tostring())
+            out.write(s.astype(float).tostring())
+            out.write(r.astype(float).tostring())
             out.close()
 
     def loadsvd(self,filename):
@@ -87,9 +87,9 @@ class factors:
         out=open(filename,"rb")
         dims=struct.unpack("lllll",out.read(struct.calcsize("lllll")))
         print(dims,8*dims[0]*8*dims[1])
-        l=np.fromstring(out.read(8*dims[0]*dims[1]),np.float)
-        s=np.fromstring(out.read(8*dims[2]),np.float)
-        r=np.fromstring(out.read(8*dims[3]*dims[4]),np.float)
+        l=np.fromstring(out.read(8*dims[0]*dims[1]),float)
+        s=np.fromstring(out.read(8*dims[2]),float)
+        r=np.fromstring(out.read(8*dims[3]*dims[4]),float)
         print(l.shape,s.shape,r.shape,dims)
         l=np.reshape(l,(dims[0],dims[1]))
         r=np.reshape(r,(dims[3],dims[4]))
@@ -122,8 +122,8 @@ class factors:
         out.write(struct.pack("lll",self.x.shape[0],
                                     self.obsdata.shape[0],
                                     self.obsdata.shape[1]))
-        out.write(self.x.astype(np.float).tostring())
-        out.write(self.obsdata.astype(np.float).tostring())
+        out.write(self.x.astype(float).tostring())
+        out.write(self.obsdata.astype(float).tostring())
         out.close()
 
     def readobsdata(self,filename):
@@ -132,8 +132,8 @@ class factors:
         sizes=struct.unpack("lll",infile.read(struct.calcsize("lll")))
         # Type is Float therefore 8 bytes per item
         print(sizes)
-        self.x=np.fromstring(infile.read(sizes[0]*8),np.float)
-        self.obsdata=np.fromstring(infile.read(8*sizes[1]*sizes[2]),np.float)
+        self.x=np.fromstring(infile.read(sizes[0]*8),float)
+        self.obsdata=np.fromstring(infile.read(8*sizes[1]*sizes[2]),float)
         print(self.obsdata.shape)
         self.obsdata=np.reshape(self.obsdata,(sizes[1],sizes[2]))
         print(self.x.shape,self.obsdata.shape)
