+++ sandbox/matchquick.py
@@ -48,8 +48,8 @@ if len(sys.argv)>=9:
 
 
 
-da=np.zeros( (len(g1l),len(g2l)), np.float)
-dt2=np.zeros( (len(g1l),len(g2l)), np.float)
+da=np.zeros( (len(g1l),len(g2l)), float)
+dt2=np.zeros( (len(g1l),len(g2l)), float)
 
 for g in g1l + g2l:
     g.u = xfab.tools.ubi_to_u_b(g.ubi)[0]
